% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lead_sheets.R
\name{clean_lead_sheets}
\alias{clean_lead_sheets}
\title{Tidy lead sheet file content into a list of tibbles.}
\usage{
clean_lead_sheets(files = NULL)
}
\arguments{
\item{files}{A vector of file paths to lead sheet excel files.}
}
\description{
test test one, two three
}
