% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nir.R
\name{clean_nir}
\alias{clean_nir}
\title{Tidy NIR csv exports into a list of tibbles.}
\usage{
clean_nir(files = NULL, sepChar = NULL, codeCols = NULL)
}
\arguments{
\item{files}{A vector of file paths to nir csv export files.
\value{
A named list of tibbles.
}}
}
\description{
A function to read in and tidy the nir machine csv output file(s)
}
