% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_test_weight.R
\name{clean_test_weight}
\alias{clean_test_weight}
\title{A function to clean GAC test weight exports}
\usage{
clean_test_weight(files = NULL)
}
\arguments{
\item{files}{Filepaths to the test weight files can be either .csv or .xlsx files}
}
\value{
A tibble of cleaned test weight data
}
\description{
This function takes exports from the GAC test weight machine and cleans them
to a tidy tibble format.
}
