% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_yield_files.R
\name{clean_yield_files}
\alias{clean_yield_files}
\title{Standardize yield file formats and merge to a single dataframe}
\usage{
clean_yield_files(files = NULL)
}
\arguments{
\item{files}{A vector of filepaths to the yield files to be standardized/merged}
}
\description{
This function takes takes the individual yield trial workbooks for a season and
standardizes their formats and merges them to a single dataframe. Often there are minor issues
that prevent the workbooks from being merged directly. This most involves discrepancies between the
column names and column types used for the variables that are otherwise the same across the worksheets.
For example, one worksheet might have the maturity date encoded as a character in a column named "MD"
while another has it encoded as a numeric in a column named "MAT". In this example, the columns would both be renamed
to a numeric named "md". This function attempts to solve other similar conflicts and then row binds all the dataframes.
}
