% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_13_pct.R
\name{convert_to_13_pct}
\alias{convert_to_13_pct}
\title{Convert to 13\% moisture content}
\usage{
convert_to_13_pct(dry_basis_pct = NULL)
}
\arguments{
\item{dry_basis_pct}{The value of some observation, typically seed protein or oil
content on a dry basis.}
}
\description{
This function just takes a value and multiplies it by 0.87. This is used
to convert protein and oil measurements from a dry b
}
