% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_season_directory.R
\name{create_season_directory}
\alias{create_season_directory}
\title{Create a new directory to hold all the files for a new breeding season}
\usage{
create_season_directory(startdir = NULL, season_name = NULL)
}
\arguments{
\item{startdir}{A file path to the directory where the new directory will be created in}

\item{season_name}{The name of the season that will be used to name the uppermost directory folder}
}
\description{
This function creates a set of nested folders to hold the data for a breeding season. The structure is set up to match
what has been used in the past and to work with the output folders that are requested by the yield file macro. The main folders
created store data for the planting, crossing, harvesting/processing, and NIR data collection stages of the breeding program.
}
\examples{
\code{create_season_directory(startdir = "C:/breeding program directory", season_name = "2021")}
}
