% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nir_file.R
\name{make_nir_file}
\alias{make_nir_file}
\title{}
\usage{
make_nir_file(leadsheet_data = NULL, nir_startnumber = 1)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
