% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_yield_lookup_table.R
\name{make_yield_lookup_table}
\alias{make_yield_lookup_table}
\title{Make a lookup table to join lead sheet test names with field data test names}
\usage{
make_yield_lookup_table(cleaned_leadsheets = NULL, cleaned_yield_files = NULL)
}
\arguments{
\item{cleaned_leadsheets}{Lead sheet data as output from the clean_lead_sheets function}

\item{cleaned_yield_files}{Yield file data as output from the clean_yield_files function}
}
\description{
This function takes the output from the leadsheet and yield file cleaning functions and lets the
user create a lookup table to map the test names used in the leadsheets to those used in the
yield files. Oftentimes we use longer names in the field note files for
each test than in the lead sheets, for example a test may be listed as LU-5E in the leadsheets
but as LU-5 Early in the yield files. However, to join the data from the leadsheets to the field data, there has
to be a way to map the names to eachother. This function takes the cleaned leadsheet and yield file data
and then gives the user an interactive worksheet with the lead sheet test names filled and allows
the user to enter in their equivalent in the field note files.
}
