% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_progress_report.R
\name{yield_progress_report}
\alias{yield_progress_report}
\title{Make a progress report for yield trial data collection}
\usage{
yield_progress_report(
  yieldfiles = NULL,
  leadsheetfiles = NULL,
  yield_name_lookup = NULL,
  has_nir = FALSE,
  nirfiles = NULL
)
}
\arguments{
\item{yieldfiles}{file paths to all the yield field data files}

\item{leadsheetfiles}{file paths to all the lead sheet files}

\item{yield_name_lookup}{A lookup table to map the test names used in the leadsheets to those used in the field notes}

\item{has_nir}{Logical: has nir data been collected yet}

\item{nirfiles}{File paths to nir data. Can be left as NULL if nir has not been collected yet}
}
\description{
This function takes the paths to the leadheet files and the yield files and makes a progress
report for each test to say what data has been collected for each test out of what data still
has to be collected
}
