% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nir_files.R
\name{clean_nir_files}
\alias{clean_nir_files}
\title{A function to clean NIR export files}
\usage{
clean_nir_files(files = NULL, nir_masterfile = NULL, select_FA = FALSE)
}
\arguments{
\item{files}{A vector of file paths to nir csv export files.}

\item{nir_masterfile}{The file path the nir masterfile}

\item{select_FA}{Logical, should fatty acid columns be included in the data}
}
\description{
This function takes all the nir file exports for a given season and extracts/cleans up the measurement data from the exports. In their
original state, the exports have some header information before the measurements themselves. This function keeps just the measurement data
and cleans it up to a tidy rectangular dataframe.
}
