% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nir_files.R
\name{clean_nir_files}
\alias{clean_nir_files}
\title{A function to clean NIR export files}
\usage{
clean_nir_files(
  code_sep_character = NULL,
  code_column_names = NULL,
  files = NULL
)
}
\description{
A function to clean NIR export files
}
