% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_test_weight.R
\name{clean_test_weight}
\alias{clean_test_weight}
\title{}
\usage{
clean_test_weight(files = NULL)
}
\arguments{
\item{files}{Filepaths to the test weight files can be either .csv or .xlsx files}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
'Jay
}
