% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nir_file.R
\name{make_nir_file}
\alias{make_nir_file}
\title{Make a dataframe for printing NIR envelopes for the yield trials}
\usage{
make_nir_file(
  leadsheet_data = NULL,
  yield_data = NULL,
  yield_lookup = NULL,
  nir_startnumber = 1
)
}
\arguments{
\item{leadsheet_data}{Leadsheet data that has already been processed with the "clean_lead_sheets" function}

\item{yield_data}{Yield file data that has already been processed with the clean_yield_files function}

\item{yield_lookup}{A lookup table used to map the test names used in the leadsheets to the test names
used in the yield files. This should be a dataframe or tibble where the first column holds the test names that
are used in the leadsheets and the second column uses the names that are used in the field note files.}

\item{nir_startnumber}{What NIR number to start on}
}
\description{
This function takes cleaned leadsheet data, cleaned yield data, and a lookup table to convert leadsheet test names
to yield file test names and returns a dataframe ready to export for printing NIR envelopes.
}
